Atualização código

+++ carStand/src/daos/PurchasedCarDAO.java
@@ -27,7 +27,7 @@ public class PurchasedCarDAO {
 				String marca = rs.getString("Car_Marca");
 				String modelo = rs.getString("Car_Modelo");
 				int cilindrada = rs.getInt("Car_Cilindrada");
-				double preco = rs.getDouble("Car_Preço");
+				int preco = rs.getInt("Car_Preço");
 				int ano = rs.getInt("Car_Ano");
 				int ID_Car= rs.getInt("ID_Car");
 				String combustivel = rs.getString("Car_Combustivel");
